backend_llm_api/requirements.txt
flask
requests
huggingface_hub
python-dotenv
